Add QR decomponsition and Orthogonal initializer
With the Orthogonal initializer in place, we no longer have to work
around the lack of it by providing other recurrent initializers for
RNNs.

* New backend functions:
  - qr
  - sign

These functions ought to be moved to the core later, but right now
they don't fully conform to TensorFlow's `tf.qr` (e.g., `qr`
takes only 2D tensors). We can fix the inconsistencies and make the
move later.