FEATURE:Support compilation with strictNullChecks
Currently, Sequential is not a valid subclass of Layer because its build
method has type `build(inputShape?: Shape | undefined)` whereas
Layer's  build method has type `build(inputShape: Shape | Shape[])`.
This breaks when trying to use Sequential in a project that has
--strictNullChecks turned on for the Typescript compiler because
Layer.build never expects undefined.

The most obvious fix is to make Layer.build's inputShape parameter
optional as well. That's what I did, but I'm not sure it's correct -- I
*just* started using tensorflow but ran into problems when I tried
setting strictNullChecks to true.

The other fix would be to make Sequential.build have type
`build(inputShape: Shape)` -- that is, to require the inputShape
parameter. But it doesn't even use it, so this seems wrong.