Make sure dilationRate is a number for 1D convolutions
When a model is exported from Keras, 1D convolutions contain a dilation rate as an array of length 1. Code in a number of places, however, assumes that the dilationRate is either a number (for 1d convolutions) or an array of length 2. For example, the dilationWidth (which is normally a number) becomes an array of length 1. This causes tfjs-node to fail as it passes the dilationWidth through to native code (as the native code expects dilationWidth to be a number).

The fix here is to convert a dilationRate of length 1 to a number as the layer is instantiated.