Let Sequential.fromConfig() handle non-array config objects
Previously, it was assumed that the config object passed to a
Sequential model during serialization is always an array (of
layer configs). But in TensorFlow r1.11, a change was made to
the serialization of Sequential, such that the config is now
an object with the 'layers' field.

This PR makes sure that TensorFlow.js is able to load such
serialized Sequential models without error.