Remove usage of operation onesLike in l2Normalize
This operation mentions its first argument as an input variable but cannot derivate, so the loss function cannot derivate if the epsilonTensor is ever used as part of the maximum operation. This is undesirable. The fill() function achieves the same result without using operators, and therefore should not cause the same issue.